LayoutTests/platform/blackberry/fast/dom/Window/closure-access-after-navigation-window-expected.txt

CONSOLE MESSAGE: line 4: TypeError: Result of expression 'opener.logTitle' [undefined] is not a function.
FAIL: Timed out waiting for notifyDone to be called
Child 1
